seperated instructors function from admin controller and rep and created new controller for instructors
