+++ client/src/constants/endpoints.ts
@@ -5,13 +5,13 @@ const END_POINTS = {
     REGISTER_INSTRUCTOR:'api/auth/instructor/instructor-register',
     LOGIN_INSTRUCTOR:'api/auth/instructor/instructor-login',
     LOGIN_ADMIN:'api/auth/admin/admin-login',
-    GET_INSTRUCTOR_REQUESTS:'api/admin/instructors/view-instructor-requests',
-    ACCEPT_INSTRUCTOR_REQUESTS:'api/admin/instructors/accept-instructor-request',
-    REJECT_INSTRUCTOR_REQUESTS:'api/admin/instructors/reject-instructor-request',
-    GET_INSTRUCTORS:'api/admin/instructors/get-all-instructors',
-    BLOCK_INSTRUCTORS:'api/admin/instructors/get-all-instructors/block-instructors',
-    UNBLOCK_INSTRUCTORS:'api/admin/instructors/get-all-instructors/unblock-instructors',
-    GET_BLOCKED_INSTRUCTORS:'api/admin/instructors/get-blocked-instructors',
+    GET_INSTRUCTOR_REQUESTS:'api/instructors/view-instructor-requests',
+    ACCEPT_INSTRUCTOR_REQUESTS:'api/instructors/accept-instructor-request',
+    REJECT_INSTRUCTOR_REQUESTS:'api/instructors/reject-instructor-request',
+    GET_INSTRUCTORS:'api/instructors/get-all-instructors',
+    BLOCK_INSTRUCTORS:'api/instructors/get-all-instructors/block-instructors',
+    UNBLOCK_INSTRUCTORS:'api/instructors/get-all-instructors/unblock-instructors',
+    GET_BLOCKED_INSTRUCTORS:'api/instructors/get-blocked-instructors',
     ADD_COURSE:'api/courses/instructors/add-course',
     GET_ALL_COURSES:'api/courses/get-all-courses',
     GET_COURSE:'api/courses/get-course',
