+++ server/src/app/usecases/management/instructorManagement.ts
@@ -1,11 +1,11 @@
-import { AdminDbInterface } from '../../../repositories/adminDbRepository';
-import HttpStatusCodes from '../../../../constants/HttpStatusCodes';
-import AppError from '../../../../utils/appError';
+import HttpStatusCodes from '../../../constants/HttpStatusCodes';
+import AppError from '../../../utils/appError';
 import { SendEmailService } from '@src/frameworks/services/sendEmailService';
+import { InstructorDbInterface } from '@src/app/repositories/instructorDbRepository';
 export const getAllInstructorRequests = async (
-  adminRepository: ReturnType<AdminDbInterface>
+  instructorRepository: ReturnType<InstructorDbInterface>
 ) => {
-  const allRequests = await adminRepository.getInstructorRequests();
+  const allRequests = await instructorRepository.getInstructorRequests();
   if (allRequests.length === 0) {
     return null;
   }
@@ -14,13 +14,15 @@ export const getAllInstructorRequests = async (
 
 export const acceptInstructorRequest = async (
   instructorId: string,
-  adminRepository: ReturnType<AdminDbInterface>,
+  instructorRepository: ReturnType<InstructorDbInterface>,
   emailService: ReturnType<SendEmailService>
 ) => {
   if (!instructorId) {
     throw new AppError('Invalid instructor id', HttpStatusCodes.BAD_REQUEST);
   }
-  const response = await adminRepository.acceptInstructorRequest(instructorId);
+  const response = await instructorRepository.acceptInstructorRequest(
+    instructorId
+  );
   if (response) {
     emailService.sendEmail(
       response.email,
@@ -34,7 +36,7 @@ export const acceptInstructorRequest = async (
 export const rejectInstructorRequest = async (
   instructorId: string,
   reason: string,
-  adminRepository: ReturnType<AdminDbInterface>,
+  instructorRepository: ReturnType<InstructorDbInterface>,
   emailService: ReturnType<SendEmailService>
 ) => {
   if (!instructorId || !reason) {
@@ -43,14 +45,14 @@ export const rejectInstructorRequest = async (
       HttpStatusCodes.BAD_REQUEST
     );
   }
-  const rejected = await adminRepository.checkRejected(instructorId);
+  const rejected = await instructorRepository.checkRejected(instructorId);
   if (rejected) {
     throw new AppError(
       'Already rejected this request',
       HttpStatusCodes.CONFLICT
     );
   }
-  const response = await adminRepository.rejectInstructorRequest(
+  const response = await instructorRepository.rejectInstructorRequest(
     instructorId,
     reason
   );
@@ -65,16 +67,16 @@ export const rejectInstructorRequest = async (
 };
 
 export const getAllInstructors = async (
-  adminRepository: ReturnType<AdminDbInterface>
+  instructorRepository: ReturnType<InstructorDbInterface>
 ) => {
-  const instructors = await adminRepository.getAllInstructors();
+  const instructors = await instructorRepository.getAllInstructors();
   return instructors;
 };
 
 export const blockInstructors = async (
   instructorId: string,
   reason: string,
-  adminRepository: ReturnType<AdminDbInterface>
+  instructorRepository: ReturnType<InstructorDbInterface>
 ) => {
   if (!instructorId || !reason) {
     throw new AppError(
@@ -82,35 +84,38 @@ export const blockInstructors = async (
       HttpStatusCodes.BAD_REQUEST
     );
   }
-  const response = await adminRepository.blockInstructors(instructorId, reason);
+  const response = await instructorRepository.blockInstructors(
+    instructorId,
+    reason
+  );
   return response;
 };
 
 export const unblockInstructors = async (
   instructorId: string,
-  adminRepository: ReturnType<AdminDbInterface>
+  instructorRepository: ReturnType<InstructorDbInterface>
 ) => {
   if (!instructorId) {
     throw new AppError('Invalid instructor id', HttpStatusCodes.BAD_REQUEST);
   }
-  const response = await adminRepository.unblockInstructors(instructorId);
+  const response = await instructorRepository.unblockInstructors(instructorId);
   return response;
 };
 
 export const getBlockedInstructors = async (
-  adminRepository: ReturnType<AdminDbInterface>
+  instructorRepository: ReturnType<InstructorDbInterface>
 ) => {
-  const blockedInstructors = await adminRepository.getBlockedInstructors();
+  const blockedInstructors = await instructorRepository.getBlockedInstructors();
   return blockedInstructors;
 };
 
 export const getInstructorByIdUseCase = async (
   instructorId: string,
-  adminRepository: ReturnType<AdminDbInterface>
+  instructorRepository: ReturnType<InstructorDbInterface>
 ) => {
   if (!instructorId) {
     throw new AppError('Invalid instructor id', HttpStatusCodes.BAD_REQUEST);
   }
-  const instructor = await adminRepository.getInstructorById(instructorId);
+  const instructor = await instructorRepository.getInstructorById(instructorId);
   return instructor;
 };
