+++ server/src/frameworks/webserver/routes/admin.ts
@@ -14,44 +14,7 @@ const adminRouter = () => {
     sendEmailServiceInterface,
     sendEmailService
   );
-  //* Instructor management
-  router.get(
-    '/instructors/view-instructor-requests',
-    controller.getInstructorRequests
-  );
-
-  router.patch(
-    '/instructors/accept-instructor-request/:instructorId',
-    controller.verifyInstructor
-  );
-
-  router.put(
-    '/instructors/reject-instructor-request',
-    controller.rejectRequest
-  );
-
-  router.get('/instructors/get-all-instructors', controller.getAllInstructor);
-
-  router.patch(
-    '/instructors/get-all-instructors/block-instructors',
-    controller.blockInstructor
-  );
-
-  router.patch(
-    '/instructors/get-all-instructors/unblock-instructors/:instructorId',
-    controller.unblockInstructor
-  );
-
-  router.get(
-    '/instructors/get-blocked-instructors',
-    controller.getBlockedInstructor
-  );
-
-  router.get(
-    '/instructors/view-instructor-requests/:instructorId',
-    controller.getInstructorById
-  );
-
+ 
   return router;
 };
