+++ server/src/frameworks/webserver/routes/index.ts
@@ -2,6 +2,7 @@ import { Application } from 'express';
 import authRouter from './auth';
 import adminRouter from './admin';
 import courseRouter from './course';
+import instructorRouter from './instructor';
 import { RedisClient } from '../../../app';
 import jwtAuthMiddleware from '../middlewares/userAuth';
 import { adminRoleCheckMiddleware } from '../middlewares/roleCheckMiddleware';
@@ -14,6 +15,7 @@ const routes = (app: Application, redisClient: RedisClient) => {
   app.use('/api/admin',jwtAuthMiddleware,adminRoleCheckMiddleware,adminRouter());
   app.use('/api/courses',jwtAuthMiddleware,courseRouter())
   app.use('/api/video-streaming',videoStreamRouter())
+  app.use('/api/instructors',instructorRouter())
 
 };
